+++ __init__.py
@@ -2,5 +2,5 @@
 Root package metadata.
 """
 
-__version__ = "1.0.9-pre-release"
+__version__ = "1.0-pre-release"