+++ monitoring/__init__.py
@@ -1,3 +1,3 @@
 """Monitoring package."""
 
-__version__ = "0.9-pre-release"
+__version__ = "1.0-pre-release"